"Fix" visibility ordering.
Using transactions for batching means that rows may end up with an id ordering
that differs from the order in which they become visible. This means that
there's a risk that consumers may miss messages from a long running batch when
shorter ones have already committed.

We "work aroud" this by just hoping that autocommit means we have reduced
window for race conditions.