Remove history around a bytea key column.
Having this present means that clients re-capitulate the schema history of
these tables each time we start up, and so other clients may observe a logs
table that requies `key` to be not-null, but doesn't have the compatability
triggers in place, and thus may fail constraint checks on insertion.

In the absence of bothering to build something marginally more sophisticated,
we can remove the mentions of `key` from the schema setup.